requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-SocketIO==5.3.5
Flask-Migrate==4.0.5
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Werkzeug==3.0.1
alembic==1.13.1
gunicorn==21.2.0
eventlet==0.33.3
Pillow==10.1.0
email-validator==2.1.0

